+++ app/Views/ProjectDocuments/list.php
@@ -13,7 +13,8 @@
         <tr>
           <th scope="col">#</th>
           <th scope="col">Project</th>
-          <th scope="col">Type</th>
+          <th scope="col">Title</th>
+          <!-- <th scope="col">Type</th> -->
           <th scope="col">Status</th>
           <th scope="col">Update Date</th>
           <th scope="col">Actions</th>
@@ -24,16 +25,19 @@
             <tr scope="row" id="<?php echo $row['id'];?>">
                 <td><?php echo $key+1; ?></td>
                 <td><?php echo $projects[$row['project-id']];?></td>
-                <td><?php echo $row['type'];?></td>
+                <td><?php  echo $row['json-object'][$row['type']]['cp-line3'];?></td>
+                <!-- <td><?php //echo $templates[$row['type']]['name'];?></td> -->
                 <td><?php echo $row['status'];?></td>
                 <td><?php echo $row['update-date'];?></td>
                 <td>
                     <a href="/documents/add/<?php echo $row['type']."/".$row['id'];?>" class="btn btn-warning">
                         <i class="fa fa-edit"></i>
                     </a>
+                  <?php if (session()->get('is-admin')): ?>
                     <a onclick="deletePlanDocument(<?php echo $row['id'];?>)" class="btn btn-danger ml-2">
                         <i class="fa fa-trash text-light"></i>
                     </a>
+                  <?php endif; ?>
                 </td>
             </tr>
         <?php endforeach; ?>
